consolidate zustand stores
