+++ src/components/Modal/VaultStrategyModal/index.ts
