+++ src/components/Modal/VaultStrategyModal/lib/VaultStrategyModalDefinitionList.tsx
