+++ src/components/Modal/VaultStrategyModal/lib/VaultStrategyModalText.tsx
