+++ src/components/Modal/VaultStrategyModal/lib/VaultStrategyModalTokenApr.tsx
@@ -14,7 +14,7 @@ import {
   useVaultTotalApr,
 } from "@/hooks"
 
-import { VaultStrategyModalDefinitionList } from "@/components/VaultStrategyModal/lib/VaultStrategyModalDefinitionList"
+import { VaultStrategyModalDefinitionList } from "@/components/Modal/VaultStrategyModal/lib/VaultStrategyModalDefinitionList"
 
 export const VaultStrategyModalTokenApr: FC<VaultProps> = (props) => {
   const { data: poolId, ...poolIdQuery } = useVaultPoolId(props)
