+++ src/components/Modal/index.ts
@@ -1,0 +1,5 @@
+export * from "./AccountModal"
+export * from "./ConnectModal"
+export * from "./ConsentModal"
+export * from "./TokenSelectModal"
+export * from "./VaultStrategyModal"
