--- src/components/VaultStrategyModal/lib/index.ts
@@ -1,4 +1,0 @@
-export * from "./VaultStrategyModalAmmApr"
-export * from "./VaultStrategyModalDefinitionList"
-export * from "./VaultStrategyModalText"
-export * from "./VaultStrategyModalTokenApr"
