+++ src/hooks/usePreviewDeposit.ts
@@ -7,7 +7,7 @@ import {
 import { queryKeys } from "@/lib/helpers"
 import { PreviewTransactionArgs } from "@/hooks/lib/api/types"
 
-import { useTxSettings } from "@/store/txSettings"
+import { useGlobalStore } from "@/store"
 
 export function usePreviewDeposit({
   enabled = true,
@@ -19,7 +19,7 @@ export function usePreviewDeposit({
   const args = {
     ...rest,
     // we store slippage as a fraction of 100; api expects slippage as a fraction of 1
-    slippage: useTxSettings((store) => store.slippageTolerance) / 100,
+    slippage: useGlobalStore((store) => store.slippageTolerance) / 100,
   }
   return useQuery({
     ...queryKeys.vaults.previewDeposit(args),
